LayoutTests/sputnik/Conformance/12_Statement/12.5_The_if_Statement/S12.5_A9_T3-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unexpected token '('
S12.5_A9_T3

PASS Expected parsing failure

TEST COMPLETE
